+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Updated Predis to v1.1.10
 - Support `composer/installers` v1 and v2
 - Link to settings page when foreign drop-in was found
+- Added `wp_cache_delete_multiple()` method
 
 ## 2.0.23